+++ src/operator.jl
@@ -81,6 +81,13 @@ transpose(op::Operator) = TransposedOperator(op)
 defaultquadstrat(lc::LinearCombinationOfOperators, tfs, bfs) =
     [defaultquadstrat(op,tfs,bfs) for op in lc.ops]
 
+defaultquadstrat(lc::LinearCombinationOfOperators, tfs::DirectProductSpace, bfs::Space) =
+    [defaultquadstrat(op,tfs.factors[1],bfs) for op in lc.ops]
+defaultquadstrat(lc::LinearCombinationOfOperators, tfs::Space, bfs::DirectProductSpace) =
+    [defaultquadstrat(op,tfs,bfs.factors[1]) for op in lc.ops]
+defaultquadstrat(lc::LinearCombinationOfOperators, tfs::DirectProductSpace, bfs::DirectProductSpace) =
+    [defaultquadstrat(op,tfs.factors[1],bfs.factors[1]) for op in lc.ops]
+
 """
     assemble(operator, test_functions, trial_functions;
         storage_policy = Val{:bandedstorage},
@@ -185,7 +192,6 @@ function assemble!(operator::Operator, test_functions::Space, trial_functions::S
     quadstrat=defaultquadstrat)
 
     quadstrat = quadstrat(operator, test_functions, trial_functions)
-
     P = Threads.nthreads()
     numchunks = P
     @assert numchunks >= 1
@@ -218,16 +224,27 @@ function assemble!(op::TransposedOperator, tfs::Space, bfs::Space, store,
 end
 
 
-function assemble!(op::LinearCombinationOfOperators, tfs::AbstractSpace, bfs::AbstractSpace,
+function assemble!(op::LinearCombinationOfOperators, tfs::Space, bfs::Space,
     store, threading = Threading{:multi};
     quadstrat=defaultquadstrat)
 
     for (a,A) in zip(op.coeffs, op.ops)
         store1(v,m,n) = store(a*v,m,n)
-        qs = quadstrat(A, tfs, bfs)
-        assemble!(A, tfs, bfs, store1, threading; quadstrat=qs)
+      #  qs = quadstrat(A, tfs, bfs)
+        assemble!(A, tfs, bfs, store1, threading; quadstrat=quadstrat)
     end
 end
+# function assemble!(op::LinearCombinationOfOperators, tfs::Space, bfs::Space,
+#     store, threading = Threading{:multi};
+#     quadstrat::Vector = defaultquadstrat)
+
+#     for (i,(a,A)) in enumerate(zip(op.coeffs, op.ops))
+#         store1(v,m,n) = store(a*v,m,n)
+#         qs = quadstrat[i]
+#         assemble!(A, tfs, bfs, store1, threading; quadstrat=qs)
+#     end
+# end
+
 
 
 # Support for direct product spaces